+++ lib/lumberjack/formatter/string_formatter.rb
@@ -2,7 +2,7 @@
 
 module Lumberjack
   class Formatter
-    # Format an object by calling +to_s+ on it.
+    # Format an object by calling `to_s` on it.
     class StringFormatter
       def call(obj)
         obj.to_s